Add support for the Rust programming language
A Rust completer is added with support for semantic completions and
GoTo*. The completer requires racerd to be built by the ycmd build.py
script. To enable racerd building, add the --racer-completer flag when
running build.py. Building racerd currently requires multirust to be
available on the system. In the future, we should check if the system
rustc is an acceptable version for building racerd.